raphtory-service/requirements.txt
# Raphtory - Temporal Graph Engine
raphtory==0.9.0

# FastAPI - Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9

# Logging
structlog==24.4.0

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx[http2]==0.27.0
